+++ Changelog.md
@@ -17,6 +17,7 @@
 - Change load `coq.ltac.*` also in commands (and not just tactics) so that
   commands can easily turn holes into goals and inhabit them calling regular
   tactics.
+- New `coq.hints.add-resolve`
 - Fix `coq.option.add` survives the end of a file
 
 ## [1.12.1] - 20-01-2021